reload configure with SIGHUP signal
Reloading the config by sending SIGHUP to the main process, instead of sending a POST request
to the /-/reload endpoint, because the latter doesn't work when the basic auth is enabled.
The password for the basic auth is encrypted using bcrypt, and I don't want to store the raw password
The SIGHUP signal is sent to the process with PID 1 in the container, which is the main process.